+++ extras/samples/box.py
@@ -1,6 +1,6 @@
+#!/usr/bin/env python
 # box.py
-from purdy.content import Code
-from purdy.tui.apps import app_factory, BoxSpec, RowSpec, simple, split
+from purdy.tui import AppFactory, BoxSpec, Code, RowSpec, TText
 
 # =============================================================================
 
@@ -12,22 +12,24 @@ def a1():
         RowSpec(2, [BoxSpec(2), BoxSpec(1)]),
     ]
 
-    app = app_factory(row_specs)
-    app.rows[0][0].content = "[blue]" + 3 * CONTENT + "[/]"
-    app.rows[0][1].content = "[green]" + 5 * CONTENT + "[/]"
-    app.rows[1][0].content = "[red]" + 7 * CONTENT + "[/]"
-    app.rows[1][1].content = "[orange]" + 7 * CONTENT + "[/]"
+    app = AppFactory.full(row_specs)
+    app.rows[0][0].append(TText("[blue]" + 3 * CONTENT + "[/]"))
+    app.rows[0][1].append(TText("[green]" + 5 * CONTENT + "[/]"))
+    app.rows[1][0].append(TText("[red]" + 7 * CONTENT + "[/]"))
+    app.rows[1][1].append(TText("[orange]" + 7 * CONTENT + "[/]"))
     return app
 
 def a2():
-    app = simple()
-    app.box.content = "[blue]" + 7 * CONTENT + "[/]"
+    app = AppFactory.simple()
+    app.box.append(TText("[blue]" + 7 * CONTENT + "[/]"))
     return app
 
 def a3():
-    app = split(20, relative_height_top=3)
-    app.top.content = "[blue]" + 7 * CONTENT + "[/]"
-    app.bottom.content = "[red]" + 7 * CONTENT + "[/]"
+    app = AppFactory.split(20, relative_height_top=3)
+    app.top.append(TText("[blue]" + 7 * CONTENT + "[/]"))
+    app.bottom.append(TText("[red]" + 7 * CONTENT + "[/]"))
     return app
 
 a1().run()
+#a2().run()
+#a3().run()